Add option to read molecule and parameters.
The molecule file is a text file that is converted to string and fed to Molecule class.
The input parameters is a yaml file which is read as a dictionary of dictionaries.
pyyaml is added to setup.py as a dependency.
A water molecule file is added to tests as a sample.
A sample parameters file is added o main directory.